Simplify controller methods overrides

+++ Web/RDD.Web/Controllers/WebController.cs
@@ -33,44 +33,12 @@ namespace RDD.Web.Controllers
         {
         }
 
-        public Task<IActionResult> PostAsync()
+        public virtual async Task<IActionResult> PostAsync()
         {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Post))
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Post))
             {
-                return ProtectedPostAsync();
+                return NotFound();
             }
-            return Task.FromResult((IActionResult)NotFound());
-        }
-
-        public Task<IActionResult> PutByIdAsync(TKey id)
-        {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
-            {
-                return ProtectedPutAsync(id);
-            }
-            return Task.FromResult((IActionResult)NotFound(id));
-        }
-
-        public Task<IActionResult> PutAsync()
-        {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
-            {
-                return ProtectedPutAsync();
-            }
-            return Task.FromResult((IActionResult)NotFound());
-        }
-
-        public Task<IActionResult> DeleteByIdAsync(TKey id)
-        {
-            if (AllowedHttpVerbs.HasVerb(HttpVerbs.Delete))
-            {
-                return ProtectedDeleteAsync(id);
-            }
-            return Task.FromResult((IActionResult)NotFound(id));
-        }
-
-        protected virtual async Task<IActionResult> ProtectedPostAsync()
-        {
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Post, false);
             ICandidate<TEntity, TKey> candidate = Helper.CreateCandidate();
 
@@ -79,8 +47,12 @@ namespace RDD.Web.Controllers
             return Ok(RDDSerializer.Serialize(entity, query));
         }
 
-        protected virtual async Task<IActionResult> ProtectedPutAsync(TKey id)
+        public virtual async Task<IActionResult> PutByIdAsync(TKey id)
         {
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
+            {
+                return NotFound();
+            }
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Put, false);
             ICandidate<TEntity, TKey> candidate = Helper.CreateCandidate();
 
@@ -94,8 +66,12 @@ namespace RDD.Web.Controllers
             return Ok(RDDSerializer.Serialize(entity, query));
         }
 
-        protected virtual async Task<IActionResult> ProtectedPutAsync()
+        public virtual async Task<IActionResult> PutAsync()
         {
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Put))
+            {
+                return NotFound();
+            }
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Put, false);
             IEnumerable<ICandidate<TEntity, TKey>> candidates = Helper.CreateCandidates();
 
@@ -111,15 +87,23 @@ namespace RDD.Web.Controllers
             return Ok(RDDSerializer.Serialize(entities, query));
         }
 
-        protected virtual async Task<IActionResult> ProtectedDeleteAsync(TKey id)
+        public virtual async Task<IActionResult> DeleteByIdAsync(TKey id)
         {
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Delete))
+            {
+                return NotFound();
+            }
             await AppController.DeleteByIdAsync(id);
 
             return Ok();
         }
 
-        protected virtual async Task<IActionResult> ProtectedDeleteAsync()
+        public virtual async Task<IActionResult> DeleteAsync()
         {
+            if (!AllowedHttpVerbs.HasVerb(HttpVerbs.Delete))
+            {
+                return NotFound();
+            }
             Query<TEntity> query = Helper.CreateQuery(HttpVerbs.Delete);
             IEnumerable<ICandidate<TEntity, TKey>> candidates = Helper.CreateCandidates();
